db: add a method to retrieve the size of an entry (#16)
Adds a new method to a billy.Database (Size) to retrieve the slot size belonging to an entry.

The purpose is if we want to evict multiple smaller items in favor of a larger one in a calling application, then we need to know how many(and which ones) to evict after inserting a new item. Without this method, the calling app would need to replicate shelf size tracking and other billy internals. With this method we can just maintain the size of items in the memory index and pick and chose from there without caring how they are managed by billy.